+++ src/main/java/step3/domain/Car.java
@@ -1,4 +1,4 @@
-package step3.service;
+package step3.domain;
 
 import step3.view.result.CarMovementResult;
