+++ src/test/java/step3/CarsTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.service.Cars;
+import step3.domain.Cars;
 
 import static org.assertj.core.api.Assertions.assertThat;
